SPAP-25398 Endret overskytende prosent til Bigdecimal

+++ src/main/java/no/skatteetaten/fastsetting/formueinntekt/forskudd/trekkrutine2024/HeleTabellen.java
@@ -1,12 +1,13 @@
 package no.skatteetaten.fastsetting.formueinntekt.forskudd.trekkrutine2024;
 
+import java.math.BigDecimal;
 import java.util.LinkedHashMap;
 
 public class HeleTabellen {
     public final LinkedHashMap<Long, Long> alleTrekk;
-    public final double overskytendeProsent;
+    public final BigDecimal overskytendeProsent;
 
-    public HeleTabellen(LinkedHashMap<Long, Long> alleTrekk, double overskytendeProsent) {
+    public HeleTabellen(LinkedHashMap<Long, Long> alleTrekk, BigDecimal overskytendeProsent) {
         this.alleTrekk = alleTrekk;
         this.overskytendeProsent = overskytendeProsent;
     }
